idの型をintに修正

+++ ReactWithASP.Server/Application/ITodoApplication.cs
@@ -5,7 +5,7 @@ namespace ReactWithASP.Server.Application;
 public interface ITodoApplication
 {
   Task<IReadOnlyList<TodoItem>> GetTodos();
-  Task<TodoItem?> GetTodo(long id);
+  Task<TodoItem?> GetTodo(int id);
   Task<TodoItem> CreateTodo(CreateTodoRequest request);
   Task<TodoItem?> UpdateTodo(
     int id,
