+++ ReactWithASP.Server/Repositories/ITodoRepository.cs
@@ -5,7 +5,7 @@ namespace ReactWithASP.Server.Repositories;
 public interface ITodoRepository
 {
   Task<IReadOnlyList<TodoItem>> GetTodos();
-  Task<TodoItem?> GetTodo(long id);
+  Task<TodoItem?> GetTodo(int id);
   Task<TodoItem> CreateTodo(TodoItem todo);
   Task<TodoItem?> UpdateTodo(
     int id,
